project/consumer/requirements.txt
kafka-python==2.0.2
pyspark==3.2.1
psycopg2-binary==2.8.6
